requirements.txt
beautifulsoup4==4.6.0
Django==2.0.7
Flask-Ask==0.9.8
Flask-RESTful==0.3.6
gtfs-realtime-bindings==0.0.5
gunicorn==19.9.0
records==0.5.2
requests==2.19.1
Unidecode==1.0.22
wheel==0.31.1
cryptogrpahy<2.2